requirements.txt
pandas >= 1.2.2
opencv-python >= 4.1
albumentations >= 0.5.2
torch >= 1.9    # using torch.inference_mode()
torchvision >= 0.8.2
matplotlib >= 3.3.4
scikit-learn >= 0.24.1
openslide-python >= 1.1.2  #openslide.lowlevel.OpenSlideError: Not a JPEG file: starts with 0xff 0x13
segmentation_models_pytorch >= 0.2
torch-optimizer >= 0.2
tqdm >= 4.0
munch >=1.16  # dict dot operator

# ray >=1.12    #I use Pytorch DDP.   distributed data parallel. using ray is more easier than using pytorch native DDP.
# tiffslide >= 1.2

#Histomics TK
# numba >= 0.54
# tifffile == 2022.5.4

#whole slide image ASAP

#  sudo apt-get install python3-dev default-libmysqlclient-dev build-essential
#pip install mysqlclient

# conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch

#pip install -i https://mirrors.aliyun.com/pypi/simple/   torch torchvision
# pip install medcam
# pip install -i https://mirrors.aliyun.com/pypi/simple/  pandas scikit-learn mysqlclient openpyxl xlwt  pydicom SimpleITK matplotlib  seaborn imgaug  albumentations  torchio captum pretrainedmodels tensorboard flask biopython python-docx
# pip install torch_optimizer timm  segmentation_models_pytorch
#pip install imagecodecs  opencv-python
# pip install -i https://mirrors.aliyun.com/pypi/simple/ openslide-python
# pip install staintools statsmodels
# pip install bs4

# sudo apt-get install openslide-tools g++ gcc libblas-dev liblapack-dev
# conda install openjdk==8.0.152
# pip install pathml

# python -m pip install detectron2 -f \
#  https://dl.fbaipublicfiles.com/detectron2/wheels/cu110/torch1.7/index.html


# conda install cocotools  #pip error


